requirements.txt
flask>=2.0.0
ultralytics>=8.0.0
pillow>=8.0.0
opencv-python>=4.5.0
numpy>=1.19.0
pytesseract>=0.3.8
